requirements.txt
PyPDF2<3.0
google-generativeai
python-dotenv
pdf2image